pyproject.toml
[project]
name = "sunchat"
version = "0.1.0"
description = "FastAPI WebSocket chat with streaming LLM, short-term memory, and emotion state"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "characterengine",
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.0.0",
    "langgraph>=1.1.6",
    "openai>=2.32.0",
    "pydantic>=2.13.2",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.44.0",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
]

[tool.uv.sources]
characterengine = { path = "characterengine", editable = true }

[project.scripts]
sunchat = "sunchat:main"

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = ["sunchat"]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]
